+++ DuoArtifacts/Artifacts/BooksDizzyArtifact.cs
@@ -1,33 +1,16 @@
-﻿using CobaltCoreModding.Definitions.ExternalItems;
-using CobaltCoreModding.Definitions.ModContactPoints;
-using HarmonyLib;
-using Microsoft.Extensions.Logging;
-using Nanoray.Shrike;
-using Nanoray.Shrike.Harmony;
-using Nickel;
-using Shockah.Shared;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Reflection.Emit;
+using HarmonyLib;
+using Nickel;
+using Shockah.Kokoro;
 
 namespace Shockah.DuoArtifacts;
 
-internal sealed class BooksDizzyArtifact : DuoArtifact
+internal sealed class BooksDizzyArtifact : DuoArtifact, IKokoroApi.IV2.IActionCostsApi.IHook
 {
-	private static ExternalSprite ShieldCostSprite = null!;
-
 	private static readonly Stack<(int Shield, int MaxShield, int Shards, int MaxShards)> OldStatusStates = new();
 	private static int TemporarilyAppliedShieldToShardsCounter;
-	private static int RenderActionShardAvailable;
-
-#pragma warning disable CS0649 // Field is never assigned to, and will always have its default value
-	[SuppressMessage("Style", "IDE0044:Add readonly modifier", Justification = "Set via IL")]
-	private static int ShardCostIconIndex;
-#pragma warning restore CS0649 // Field is never assigned to, and will always have its default value
 
 	public int ShardsToRestoreOnNextCombat;
 
@@ -43,23 +26,6 @@ internal sealed class BooksDizzyArtifact : DuoArtifact
 			original: AccessTools.DeclaredMethod(typeof(Combat), nameof(Combat.DrainCardActions)),
 			prefix: new HarmonyMethod(GetType(), nameof(Combat_DrainCardActions_Prefix)),
 			finalizer: new HarmonyMethod(GetType(), nameof(Combat_DrainCardActions_Finalizer))
-		);
-		harmony.Patch(
-			original: AccessTools.DeclaredMethod(typeof(Card), nameof(Card.MakeAllActionIcons)),
-			transpiler: new HarmonyMethod(GetType(), nameof(Card_MakeAllActionIcons_Transpiler))
-		);
-		harmony.Patch(
-			original: AccessTools.DeclaredMethod(typeof(Card), nameof(Card.RenderAction)),
-			prefix: new HarmonyMethod(GetType(), nameof(Card_RenderAction_Prefix)),
-			transpiler: new HarmonyMethod(GetType(), nameof(Card_RenderAction_Transpiler))
-		);
-		harmony.Patch(
-			original: typeof(Card).GetMethods(AccessTools.all).First(m => m.Name.StartsWith("<RenderAction>g__ShardcostIcon") && m.ReturnType == typeof(void)),
-			transpiler: new HarmonyMethod(GetType(), nameof(Card_RenderAction_ShardcostIcon_Transpiler))
-		);
-		harmony.Patch(
-			original: AccessTools.DeclaredMethod(typeof(AVariableHint), nameof(AVariableHint.GetTooltips)),
-			transpiler: new HarmonyMethod(GetType(), nameof(AVariableHint_GetTooltips_Transpiler))
 		);
 		harmony.Patch(
 			original: AccessTools.DeclaredMethod(typeof(ShieldGun), "GetShieldAmt"),
@@ -88,15 +54,6 @@ internal sealed class BooksDizzyArtifact : DuoArtifact
 		harmony.Patch(
 			original: AccessTools.DeclaredMethod(typeof(Glimmershot), nameof(Glimmershot.GetActions)),
 			postfix: new HarmonyMethod(GetType(), nameof(Glimmershot_GetActions_Postfix))
-		);
-	}
-
-	protected internal override void RegisterArt(ISpriteRegistry registry, string namePrefix, DuoArtifactDefinition definition)
-	{
-		base.RegisterArt(registry, namePrefix, definition);
-		ShieldCostSprite = registry.RegisterArtOrThrow(
-			id: $"{typeof(ModEntry).Namespace}.Icon.ShieldCost",
-			file: new FileInfo(Path.Combine(Instance.ModRootFolder!.FullName, "assets", "Icons", "ShieldCost.png"))
 		);
 	}
 
@@ -175,8 +132,7 @@ internal sealed class BooksDizzyArtifact : DuoArtifact
 	{
 		if (__instance != s.ship)
 			return;
-		var artifact = s.EnumerateAllArtifacts().FirstOrDefault(a => a is BooksDizzyArtifact);
-		if (artifact is null)
+		if (s.EnumerateAllArtifacts().FirstOrDefault(a => a is BooksDizzyArtifact) is not { } artifact)
 			return;
 
 		var (_, oldShards) = PopStatusState();
@@ -223,8 +179,7 @@ internal sealed class BooksDizzyArtifact : DuoArtifact
 	{
 		if (TemporarilyAppliedShieldToShardsCounter <= 0)
 			return;
-		var artifact = g.state.EnumerateAllArtifacts().FirstOrDefault(a => a is BooksDizzyArtifact);
-		if (artifact is null)
+		if (g.state.EnumerateAllArtifacts().FirstOrDefault(a => a is BooksDizzyArtifact) is not { } artifact)
 			return;
 
 		TemporarilyAppliedShieldToShardsCounter--;
@@ -256,159 +211,6 @@ internal sealed class BooksDizzyArtifact : DuoArtifact
 		// honestly not sure what to do about this?
 		//if (leftOverToRemove > 0)
 		//	g.state.ship.DirectHullDamage(s, c, leftOverToRemove);
-	}
-
-	private static IEnumerable<CodeInstruction> Card_MakeAllActionIcons_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		// ReSharper disable PossibleMultipleEnumeration
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(
-					ILMatches.Ldarg(2),
-					ILMatches.Ldfld("ship"),
-					ILMatches.LdcI4((int)Status.shard),
-					ILMatches.Call("Get")
-				)
-				.Insert(
-					SequenceMatcherPastBoundsDirection.After, SequenceMatcherInsertionResultingBounds.IncludingInsertion,
-					new CodeInstruction(OpCodes.Ldarg_2),
-					new CodeInstruction(OpCodes.Call, AccessTools.DeclaredMethod(typeof(BooksDizzyArtifact), nameof(Card_MakeAllActionIcons_Transpiler_ModifyStatusValue)))
-				)
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger!.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Name, ex);
-			return instructions;
-		}
-		// ReSharper restore PossibleMultipleEnumeration
-	}
-
-	private static int Card_MakeAllActionIcons_Transpiler_ModifyStatusValue(int value, State state)
-	{
-		if (!state.EnumerateAllArtifacts().Any(a => a is BooksDizzyArtifact))
-			return value;
-		return value + state.ship.Get(Status.shield);
-	}
-
-	private static void Card_RenderAction_Prefix(int shardAvailable)
-		=> RenderActionShardAvailable = shardAvailable;
-
-	private static IEnumerable<CodeInstruction> Card_RenderAction_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		// ReSharper disable PossibleMultipleEnumeration
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(
-					ILMatches.Ldarg(4).ExtractLabels(out var labels),
-					ILMatches.Ldloc<int>(originalMethod).CreateLdlocInstruction(out var ldlocLoopIterator),
-					ILMatches.Instruction(OpCodes.Clt),
-					ILMatches.LdcI4(0),
-					ILMatches.Instruction(OpCodes.Ceq),
-					ILMatches.AnyLdloca,
-					ILMatches.Instruction(OpCodes.Call)
-				)
-				.Insert(
-					SequenceMatcherPastBoundsDirection.Before, SequenceMatcherInsertionResultingBounds.JustInsertion,
-					ldlocLoopIterator.Value.WithLabels(labels),
-					new CodeInstruction(OpCodes.Stsfld, AccessTools.DeclaredField(typeof(BooksDizzyArtifact), nameof(ShardCostIconIndex)))
-				)
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger!.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Name, ex);
-			return instructions;
-		}
-		// ReSharper restore PossibleMultipleEnumeration
-	}
-
-	private static IEnumerable<CodeInstruction> Card_RenderAction_ShardcostIcon_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		// ReSharper disable PossibleMultipleEnumeration
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(
-					ILMatches.Ldarg(0),
-					ILMatches.Brtrue,
-					ILMatches.LdcI4((int)StableSpr.icons_shardcostoff),
-					ILMatches.Br,
-					ILMatches.LdcI4((int)StableSpr.icons_shardcost)
-				)
-				.Insert(
-					SequenceMatcherPastBoundsDirection.After, SequenceMatcherInsertionResultingBounds.IncludingInsertion,
-					new CodeInstruction(OpCodes.Ldarg_0),
-					new CodeInstruction(OpCodes.Call, AccessTools.DeclaredMethod(typeof(BooksDizzyArtifact), nameof(Card_RenderAction_ShardcostIcon_Transpiler_ModifySprite)))
-				)
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger!.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Name, ex);
-			return instructions;
-		}
-		// ReSharper restore PossibleMultipleEnumeration
-	}
-
-	private static int Card_RenderAction_ShardcostIcon_Transpiler_ModifySprite(int spriteId, bool costMet)
-	{
-		if (!costMet)
-			return spriteId;
-		if (MG.inst.g.state is not { } state || !state.EnumerateAllArtifacts().Any(a => a is BooksDizzyArtifact))
-			return spriteId;
-
-		var shield = state.ship.Get(Status.shield);
-		var shards = state.ship.Get(Status.shard);
-
-		var totalIndex = ShardCostIconIndex + (shards + shield - RenderActionShardAvailable);
-		if (totalIndex > shards)
-			return ShieldCostSprite.Id!.Value;
-		return spriteId;
-	}
-
-	private static IEnumerable<CodeInstruction> AVariableHint_GetTooltips_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		// ReSharper disable PossibleMultipleEnumeration
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(
-					ILMatches.Ldarg(1),
-					ILMatches.Ldfld("ship"),
-					ILMatches.Ldarg(0),
-					ILMatches.Ldflda("status"),
-					ILMatches.Call("get_Value"),
-					ILMatches.Call("Get")
-				)
-				.Insert(
-					SequenceMatcherPastBoundsDirection.After, SequenceMatcherInsertionResultingBounds.IncludingInsertion,
-					new CodeInstruction(OpCodes.Ldarg_0),
-					new CodeInstruction(OpCodes.Ldarg_1),
-					new CodeInstruction(OpCodes.Call, AccessTools.DeclaredMethod(typeof(BooksDizzyArtifact), nameof(AVariableHint_GetTooltips_Transpiler_ModifyStatusValue)))
-				)
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger!.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Name, ex);
-			return instructions;
-		}
-		// ReSharper restore PossibleMultipleEnumeration
-	}
-
-	private static int AVariableHint_GetTooltips_Transpiler_ModifyStatusValue(int value, AVariableHint hint, State state)
-	{
-		if (!state.EnumerateAllArtifacts().Any(a => a is BooksDizzyArtifact))
-			return value;
-		return hint.status switch
-		{
-			Status.shield => value + state.ship.Get(Status.shard),
-			Status.shard => value + state.ship.Get(Status.shield),
-			_ => value
-		};
 	}
 
 	private static void ShieldGun_GetShieldAmt_Postfix(State s, ref int __result)
@@ -478,5 +280,32 @@ internal sealed class BooksDizzyArtifact : DuoArtifact
 			mode = AStatusMode.Set,
 			targetPlayer = true
 		}).AsCardAction);
+	}
+	
+	public bool ModifyActionCost(IKokoroApi.IV2.IActionCostsApi.IHook.IModifyActionCostArgs args)
+	{
+		HandleAnyCost(args.Cost);
+		return false;
+
+		void HandleAnyCost(IKokoroApi.IV2.IActionCostsApi.ICost cost)
+		{
+			if (ModEntry.Instance.KokoroApi.ActionCosts.AsResourceCost(cost) is { } resourceCost)
+				HandleResourceCost(resourceCost);
+			else if (ModEntry.Instance.KokoroApi.ActionCosts.AsCombinedCost(cost) is { } combinedCost)
+				HandleCombinedCost(combinedCost);
+		}
+
+		void HandleCombinedCost(IKokoroApi.IV2.IActionCostsApi.ICombinedCost combinedCost)
+		{
+			foreach (var cost in combinedCost.Costs)
+				HandleAnyCost(cost);
+		}
+		
+		void HandleResourceCost(IKokoroApi.IV2.IActionCostsApi.IResourceCost cost)
+		{
+			if (!cost.PotentialResources.Any(r => ModEntry.Instance.KokoroApi.ActionCosts.AsStatusResource(r) is { Status: Status.shard }))
+				return;
+			cost.PotentialResources.Add(ModEntry.Instance.KokoroApi.ActionCosts.MakeStatusResource(Status.shield));
+		}
 	}
 }